pyproject.toml
[tool.poetry]
name = "slurpit-netbox"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "slurpit_netbox", from = "src"}]


[tool.poetry.dependencies]
python = "^3.10"
netmiko = "^4.2.0"
requests = "^2.31.0"
arrow = "^1.3.0"


[tool.poetry.group.develop.dependencies]
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-tables2 = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
